api/requirements.txt
PyGithub==1.54.1
expertai-nlapi==2.1.3
Flask==1.1.2
nltk==3.5
numpy==1.20.1
gunicorn